Push criterion to separate directory
In order to maintain MRSV
