+++ compare/benches/bench_bits.rs
@@ -52,16 +52,16 @@ fn bitting(c: &mut Criterion) {
                             if len < runs {
                                 len = bitter.refill_lookahead();
                             }
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
                             len -= runs;
@@ -74,13 +74,13 @@ fn bitting(c: &mut Criterion) {
                             if len < runs {
                                 len = bitter.refill_lookahead();
                             }
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
-            
+
                             black_box(bitter.peek(x));
                             bitter.consume(x);
-            
+
                             len -= runs;
                         }
                     }
@@ -96,13 +96,13 @@ fn bitting(c: &mut Criterion) {
                             bitter.refill_lookahead();
                             let lo = bitter.peek(bitter::MAX_READ_BITS);
                             bitter.consume(bitter::MAX_READ_BITS);
-            
+
                             let hi_bits = x - bitter::MAX_READ_BITS;
                             bitter.refill_lookahead();
-            
+
                             let hi = bitter.peek(hi_bits);
                             bitter.consume(hi_bits);
-            
+
                             black_box((hi << bitter::MAX_READ_BITS) + lo);
                         }
                     }
@@ -258,6 +258,208 @@ fn bitting(c: &mut Criterion) {
     group.finish();
 }
 
-criterion_group!(benches, bitting,);
+fn real_world1(c: &mut Criterion) {
+    // Parse a rocket league Quaternion, which is bit reads of 2 + 18 + 18 + 18 = 56;
+    let mut group = c.benchmark_group("real-world-1");
+
+    group.throughput(Throughput::Bytes((56 as u64 * ITER) / 8));
+
+    group.bench_function("bitter-auto", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                black_box(bits.read_bits(2));
+                black_box(bits.read_bits(18));
+                black_box(bits.read_bits(18));
+                black_box(bits.read_bits(18));
+            }
+        })
+    });
+
+    group.bench_function("bitter-manual", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                let len = bits.refill_lookahead();
+                assert!(len >= bitter::MAX_READ_BITS);
+
+                black_box(bits.peek(2));
+                bits.consume(2);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+            }
+        })
+    });
+
+    group.bench_function("bitter-unchecked", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                unsafe { bits.refill_lookahead_unchecked() };
+
+                black_box(bits.peek(2));
+                bits.consume(2);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+
+                black_box(bits.peek(18));
+                bits.consume(18);
+            }
+        })
+    });
+
+    group.finish();
+}
+
+fn real_world2(c: &mut Criterion) {
+    // Parse a rocket league Vector3i (simiplified)
+    let mut group = c.benchmark_group("real-world-2");
+
+    group.throughput(Throughput::Bytes((62 as u64 * ITER) / 8));
+
+    group.bench_function("bitter-auto", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                black_box(bits.read_bits(4));
+                black_box(bits.read_bits(1));
+                black_box(bits.read_bits(19));
+                black_box(bits.read_bits(19));
+                black_box(bits.read_bits(19));
+            }
+        })
+    });
+
+    group.bench_function("bitter-manual", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                let len = bits.refill_lookahead();
+                assert!(len >= 24);
+                black_box(bits.peek(4));
+                bits.consume(4);
+
+                black_box(bits.peek(1));
+                bits.consume(1);
+
+                black_box(bits.peek(19));
+                bits.consume(19);
+
+                let len = bits.refill_lookahead();
+                assert!(len > 36);
+                black_box(bits.peek(19));
+                bits.consume(19);
+
+                black_box(bits.peek(19));
+                bits.consume(19);
+            }
+        })
+    });
+
+    group.bench_function("bitter-unchecked", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA);
+            for _ in 0..ITER {
+                unsafe { bits.refill_lookahead_unchecked() };
+
+                black_box(bits.peek(4));
+                bits.consume(4);
+
+                black_box(bits.peek(1));
+                bits.consume(1);
+
+                black_box(bits.peek(19));
+                bits.consume(19);
+
+                unsafe { bits.refill_lookahead_unchecked() };
+
+                black_box(bits.peek(19));
+                bits.consume(19);
+
+                black_box(bits.peek(19));
+                bits.consume(19);
+            }
+        })
+    });
+
+    group.finish();
+}
+
+fn read_bytes(c: &mut Criterion) {
+    let mut group = c.benchmark_group("read_bytes");
+
+    group.bench_function("aligned", |b| {
+        b.iter(|| {
+            let mut buf = [0u8; 7];
+            let mut bitter = LittleEndianReader::new(&DATA[..]);
+            for _ in 0..ITER {
+                black_box(bitter.read_bytes(&mut buf));
+            }
+        })
+    });
+
+    group.bench_function("unaligned", |b| {
+        b.iter(|| {
+            let mut buf = [0u8; 7];
+            let mut bitter = LittleEndianReader::new(&DATA[..]);
+            bitter.read_bit();
+            for _ in 0..ITER {
+                black_box(bitter.read_bytes(&mut buf));
+            }
+        })
+    });
+
+    group.finish();
+}
+
+fn signed(c: &mut Criterion) {
+    let mut group = c.benchmark_group("signed_reads");
+
+    let bits_to_read = 33;
+    group.throughput(Throughput::Bytes((bits_to_read as u64 * ITER) / 8));
+
+    group.bench_function("auto", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA[..]);
+            for _ in 0..ITER {
+                black_box(bits.read_signed_bits(bits_to_read));
+            }
+        })
+    });
+
+    group.bench_function("manual", |b| {
+        b.iter(|| {
+            let mut bits = LittleEndianReader::new(&DATA[..]);
+            for _ in 0..ITER {
+                let _len = bits.refill_lookahead();
+                let val = bits.peek(bits_to_read);
+                bits.consume(bits_to_read);
+                black_box(bitter::sign_extend(val, bits_to_read));
+            }
+        })
+    });
+
+    group.finish();
+}
+
+criterion_group!(
+    benches,
+    bitting,
+    real_world1,
+    real_world2,
+    read_bytes,
+    signed
+);
 
 criterion_main!(benches);
